+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 0.0.7
 
+- Fix: differentiate between Identifier and TypeName nodes
 - Fix: don't allow illegal ref cycles
-- Fix various bugs in the parser/checker
+- Fix various other small bugs in the parser/checker
 - Massively improved unit test foundation
 - Don't retain comments as part of the AST (might be added back in later)
 - Fix for non-node envs (missing `process` global)